Fix attribute assignment of nil
Add specs

Extract shared examples

Remove unnecessary files addition

Remove unnecessary require
