+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
 require 'bundler/setup'
 require 'valued'
+require 'valued/shared_examples/valued_shared_examples'
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
